+++ tests/full/test_full.tf
@@ -29,7 +29,7 @@ module "main" {
     credential_policy = "CP1"
     dvs_version       = "6.5"
     statistics        = true
-    mgmt_epg          = "oob"
+    mgmt_epg_type     = "oob"
   }]
   credential_policies = [{
     name     = "CP1"